(refactor) moved markets' components to markets'  folder
