+++ app/src/components/common/index.ts
@@ -44,9 +44,6 @@ export { ThreeBoxComments } from './three_box_comments'
 export { TitleValue } from './title_value'
 export { Tokens } from './tokens'
 export { Tooltip } from './tooltip'
-export { TransactionDetailsCard } from './transaction_details_card'
-export { TransactionDetailsLine } from './transaction_details_line'
-export { TransactionDetailsRow } from './transaction_details_row'
 export { ViewCard } from './view_card'
 export { WalletBalance } from './wallet_balance'
 export { Well } from './well'
