+++ app/src/components/market/advanced_filters/img/currency/index.ts
