+++ app/src/components/market/market_buy.tsx
@@ -28,18 +28,18 @@ import {
   SubsectionTitleWrapper,
   TextfieldCustomPlaceholder,
   TitleValue,
-  TransactionDetailsCard,
-  TransactionDetailsLine,
-  TransactionDetailsRow,
   ViewCard,
   WalletBalance,
 } from '../common'
 import { BigNumberInputReturn } from '../common/big_number_input'
 import { OutcomeTable } from '../common/outcome_table'
 import { SetAllowance } from '../common/set_allowance'
-import { ValueStates } from '../common/transaction_details_row'
 import { FullLoading } from '../loading'
 import { ModalTwitterShare } from '../modal/modal_twitter_share'
+
+import { TransactionDetailsCard } from './transaction_details_card'
+import { TransactionDetailsLine } from './transaction_details_line'
+import { TransactionDetailsRow, ValueStates } from './transaction_details_row'
 
 const LeftButton = styled(Button)`
   margin-right: auto;
